pyproject.toml
[project]
name = "wardenwg"
version = "0.1.0"
description = "WireGuard 管理面 MVP"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14.0",
    "apscheduler>=3.10.4",
    "fastapi>=0.115.0",
    "jinja2>=3.1.4",
    "paramiko>=3.5.0",
    "pydantic-settings>=2.6.1",
    "python-multipart>=0.0.12",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.36",
    "uvicorn[standard]>=0.32.0",
    "cryptography>=44.0.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.3.3",
    "ruff>=0.8.2",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

